+++ src/commonMain/kotlin/number/PositiveInt.kt
@@ -12,6 +12,7 @@ import kotools.types.experimental.ExperimentalKotoolsTypesApi
 import kotools.types.experimental.ExperimentalRangeApi
 import kotools.types.experimental.NotEmptyRange
 import kotools.types.experimental.notEmptyRangeOf
+import kotools.types.experimental.range
 import kotools.types.internal.ExperimentalSince
 import kotools.types.internal.KotoolsTypesVersion
 import kotools.types.internal.Since